[#20] fix: ReportMapSummaryResponse dto scamCnt 타입 수정

+++ BE/src/main/java/com/swkim/safetrip/dto/response/ReportMapSummaryResponse.java
@@ -14,7 +14,7 @@ public class ReportMapSummaryResponse{
     private Long id;
     private String name;
     private String type;
-    private Integer scamCnt;
+    private Long scamCnt;
     private double lat;
     private double lng;
 }
